POST Req

+++ ControllPanel/Model/AddressDTO.cs
@@ -13,7 +13,7 @@ namespace ControllPanel.Model
             public string Country { get; set; }
             public string City { get; set; }
             public string Street { get; set; }
-            public int ZipCode { get; set; }
+            public long ZipCode { get; set; }
         }  
         public class AddressDTO :CreateAddressDTO
         {
